Move User to own entity namespace
