+++ src/Entity/Project/Project.php
@@ -9,7 +9,7 @@ use App\Entity\Interface\AccountingOwnerInterface;
 use App\Entity\Trait\MigratedEntity;
 use App\Entity\Trait\TimestampedCreationEntity;
 use App\Entity\Trait\TimestampedUpdationEntity;
-use App\Entity\User;
+use App\Entity\User\User;
 use App\Repository\Project\ProjectRepository;
 use Doctrine\ORM\Mapping as ORM;
